Rename index.js to server.js
Also had to tweak a test in contactSearch.route.test.js as this was causing the server tests to fail if they ran later in the order.
