+++ test/routes/root.route.test.js
@@ -4,7 +4,7 @@ const Lab = require('lab')
 const lab = exports.lab = Lab.script()
 const Code = require('code')
 // const DOMParser = require('xmldom').DOMParser
-const server = require('../../index')
+const server = require('../../server')
 
 const ActiveDirectoryAuthService = require('../../src/services/activeDirectoryAuth.service')
